Extract types to the types file
